requirements.txt
bcrypt==4.2.1
Pillow==12.2.0
tkcalendar==1.6.1
fpdf2==2.8.2
matplotlib==3.10.8
babel==2.16.0
